requirements.txt
Flask==3.0.3
imutils==0.5.4
opencv_python==4.9.0.80
Pillow==10.0.0
Pillow==10.3.0
skimage==0.0
Werkzeug==3.0.2
